refactor: extract ContractAddressItem and improve validation error handling
- Extract ContractAddressItem into separate file for better code organization
- Fix input focus issue by using stable array index as key instead of dynamic contractKey
- Fix heading hierarchy: change Dialog.Title from h3 to h2, replace nested h2 with div
- Prevent save when validation API calls fail by tracking errors in validation results
- Display actual error messages from validation failures instead of generic messages
- Add errorMessage field to validation results for better error reporting
